Audio/VaKit: Initial commit of VKAT, heavily work in progress [build fix]. bugref:10008
svn:sync-xref-src-repo-rev: r144259

+++ src/VBox/Devices/Audio/DrvHostAudioAlsa.cpp
@@ -1,4 +1,4 @@
-/* $Id: DrvHostAudioAlsa.cpp 88923 2021-05-07 13:34:51Z andreas.loeffler@oracle.com $ */
+/* $Id: DrvHostAudioAlsa.cpp 88924 2021-05-07 13:41:46Z andreas.loeffler@oracle.com $ */
 /** @file
  * Host audio driver - Advanced Linux Sound Architecture (ALSA).
  */
@@ -1102,7 +1102,7 @@ static DECLCALLBACK(uint32_t) drvHostAlsaAudioHA_StreamGetPending(PPDMIHOSTAUDIO
             case SND_PCM_STATE_DRAINING:
                 if (rc >= 0)
                 {
-                    if (cFramesAvail >= pStreamALSA->Cfg.Backend.cFramesBufferSize)
+                    if ((uint32_t)cFramesAvail >= pStreamALSA->Cfg.Backend.cFramesBufferSize)
                         cbPending = 0;
                     else
                         cbPending = PDMAudioPropsFramesToBytes(&pStreamALSA->Cfg.Props, cFramesDelay);
